membership: Re-structure pallet
